Improve Top 10 bookmarks list maintenance, ensuring that we look at both visit count and last visit date to keep the list stable in the face of changing visit counts, and optimize list building slightly.
Add methods to allow bookmark sorting, currently only used by the top 10 list.

Remove a nasty hack that set the parent of a bookmark item to an NSNumber for Rendezvous bookmarks, replacing it with a RendezvousBookmark Bookmark subclass.

Fix a bug in the Bookmark enumerator, that could cause it to propagate above its root folder, and hang in some cases.

Optimize the notification handlers for various bookmark changes (notably visit count changes) to reduce the amount of work done when resetting all visit counts.

Optimize the building of the dictionary that maps from bookmarks to their favicon urls.

Add a stack-based C++ class that can be used to time an event.